Add the option to dotfiles task to the local path if exists already

+++ lib/homey/tasks.rb
@@ -3,11 +3,15 @@ module Homey
     desc "dotfiles", "Fetching your dotfiles from github"
     method_option :path, type: :string
     method_option :force, aliases: "-f", type: :boolean
+    method_option :local, aliases: "-l", type: :boolean
     def dotfiles(repo)
       path = prepare_paths(options.fetch("path", "~/.dotfiles")).first
-      FileUtils.rm_rf(path) if options["force"]
-      command = "git clone git@github.com:#{repo}.git #{path}"
-      system command
+      if !options[:local]
+        FileUtils.rm_rf(path) if options["force"]
+        command = "git clone git@github.com:#{repo}.git #{path}"
+        system command
+      end
+      File.open("#{ENV['HOME']}/.homey", 'w') { |f| f.write(path) }
     end
 
     desc "create_symlinks", "Creating the sym links that you have set in home.yml"
